+++ packages/runtime-common/fetch-browser.ts
@@ -1,0 +1,4 @@
+export function createEnvironmentAwareFetch(): typeof globalThis.fetch {
+  // Browser environment - use native fetch
+  return globalThis.fetch.bind(globalThis);
+}
