+++ packages/runtime-common/virtual-network.ts
@@ -1,10 +1,5 @@
-import { RealmPaths, ensureTrailingSlash } from './paths';
-import { baseRealm, isNode } from './index';
-import {
-  registerCardReferencePrefix,
-  rri,
-  type RealmIdentifier,
-} from './card-reference-resolver';
+import { RealmPaths } from './paths';
+import { baseRealm } from './index';
 import type { ModuleDescriptor } from './package-shim-handler';
 import {
   PackageShimHandler,
@@ -13,6 +8,7 @@ import {
 } from './package-shim-handler';
 import type { Readable } from 'stream';
 import { fetcher, type FetcherMiddlewareHandler } from './fetcher';
+import { createEnvironmentAwareFetch } from '#fetch';
 
 export interface ResponseWithNodeStream extends Response {
   nodeStream?: Readable;
@@ -271,83 +267,6 @@ export function isUrlLike(moduleIdentifier: string): boolean {
     moduleIdentifier.startsWith('http://') ||
     moduleIdentifier.startsWith('https://')
   );
-}
-
-/**
- * Creates a fetch implementation that's appropriate for the current environment.
- * In Node.js, it enhances localhost subdomain resolution using Undici agent.
- * In browsers, it uses native fetch.
- */
-function createEnvironmentAwareFetch(): typeof globalThis.fetch {
-  if (!isNode) {
-    // Browser environment - use native fetch
-    return globalThis.fetch.bind(globalThis);
-  }
-
-  // Node.js environment - create enhanced fetch with Undici dispatcher
-  try {
-    // Check if we're in a browser-like environment (even if isNode is true)
-    if (
-      typeof window !== 'undefined' ||
-      typeof globalThis.fetch === 'undefined'
-    ) {
-      // Browser environment or no fetch available - use native fetch
-      return globalThis.fetch.bind(globalThis);
-    }
-
-    // Check if undici and dns are available at runtime
-    let undici: any;
-    let dns: any;
-    try {
-      undici = require('undici');
-      dns = require('dns');
-    } catch (e) {
-      // Undici not available - fallback to native fetch
-      return globalThis.fetch.bind(globalThis);
-    }
-
-    const { Agent } = undici;
-
-    // Create a custom agent with localhost subdomain resolution
-    const agent = new Agent({
-      connect: {
-        // This replaces dns.lookup for sockets created by this Agent
-        lookup(hostname: string, options: any, cb: any) {
-          if (hostname?.endsWith('.localhost')) {
-            if (options.all) {
-              // Return array format if options.all is true
-              return cb(null, [{ address: '127.0.0.1', family: 4 }], null);
-            } else {
-              // Return standard format otherwise
-              return cb(null, '127.0.0.1', 4);
-            }
-          }
-          // Use default DNS lookup for all other hostnames
-          // Use a lazy-loaded function to avoid bundler issues
-          function performDNSLookup() {
-            try {
-              return dns.lookup(hostname, options, cb);
-            } catch (e) {
-              return cb(new Error('DNS lookup failed'), null, null);
-            }
-          }
-          return performDNSLookup();
-        },
-      },
-    });
-
-    // Create a custom fetch function that uses our agent
-    return async (input: RequestInfo | URL, init?: RequestInit) => {
-      let fetch = globalThis.fetch.bind(globalThis);
-      return fetch(input, {
-        ...init,
-        dispatcher: agent,
-      } as any);
-    };
-  } catch (e) {
-    // Fallback to native fetch if undici setup fails
-    return globalThis.fetch.bind(globalThis);
-  }
 }
 
 // This is to handle a very mysterious situation in our CI environment where
